refactor(migration): Switch quote snowflake type conversion system to use Mongock
